refactor array sequence zip operators
